feat: Automate provisioning of local environments
The assignment requires to provision local environments quickly.
To simplify and speed up the provisioning of local environments,
which are neccessary for fast software iteration and debugging,
this adds a solution which automatically provisions local environments
based on Minikube and Pulumi with filesystem backend - which should
be cheap to run and easy to use.