fix: Networking in the solution diagram
NetworkPolicies are additive, so there cannot be one blocking all
ingress, and another allowing it. This changes the diagram to only
introduce one NetworkPolicy which blocks all access, but still at
the same time allows ingress to the `invoice-app`. `invoice-app`
must be accessible outside the cluster, so this adds the Ingress
onto the diagram.